Fix secret expr validation + move ExprValidator API nicer
